Update patch set 1
Patch Set 1: Code-Review+1

Marc-Andre, I'm sorry I had overlooked this review request before!

Regarding not expanding all levels, as far as I can tell this is not a limit imposed by the TreeViewer, but rather an artefact of the content provider, IBContentProvider, being an AsyncTreeContentProvider which computes the children of a node asynchronously while returning a "work in progress" node synchronously. When the expandAll() operation tries to expand a child recursively, it sometimes sees the "work in progress" node which has no children yet. I think working around this would require writing a more sophisticated version of TreeViewer.expandAll() which remembers that "work in progress" nodes need to be expanded after they're resolved, and reacts to them being resolved (I guess in refresh()) to do that.

Regarding the performance issue, it seems like every children computation requires acquiring the index lock, listing includes requires reading them from a linked list associated with each file in the index, and resolving each include requires a search in the PDOM's "file index" B-tree. I'm guessing these operations in aggregate contribute to the observed performance for a large project with many files and includes, although it would be useful to do some profiling to see which one in particular.

I think it would be reasonable to commit this in its current state, and iterate on the async-expanding and performance issues as follow-ups.

Patch-set: 1
Reviewer: Gerrit User 973 <973@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Code-Review=+1