Update patch set 1
Patch Set 1:

Re: "why delete action is not enabled in editor?", I agree that it makes sense to support it. It just requires a different implementation (operating on an ITextSelection rather than an ICElement), and I suggested to Luis that we leave that as a follow-up, to keep this first patch simple.

Patch-set: 1