Update patch set 2
Patch Set 2:

Thanks, this generally looks good!

One thing I noticed is that there is now a "Delete" menu item in the "Refactor" menu of Eclipse's main menu bar, which is never active. I believe this is because the delete action's entry in plugin.xml specifies a menubarPath, but CRefactoringActionGroup.fillActionBars() doesn't add the delete action to the action bars.

Since we're not supporting the delete action in the editor yet, we probably don't want it in the main menu bar's "Refactor" menu yet, so I would suggest removing the menubarPath attribute from plugin.xml for now. Later when we add support for invoking this action from the editor, we can add back that attribtue and add the action to CRefactoringActionGroup.fillActionBars() at the same time.

Patch-set: 2
Label: Verified=0