Update patch set 3
Patch Set 3:

This patch removes the plugin.xml entry altogether. I was thinking of just removing the menubarPath attribute.

(I tried the patch and I didn't notice anything being broken as a result of not having the plugin.xml entry, but it's probably necessary for some things other than adding the action to the menu bar. For example, if we later add a keyboard shortcut for the action, that will probably require it to have an entry in plugin.xml.)

Patch-set: 3
Reviewer: Gerrit User 973 <973@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0