Bug 481126 - Convert Acorn AST into Java
Added the class QmlASTNodeHandler that creates and manages Proxy
instances of the QML AST node interfaces.  The proxies are created
dynamically based on what methods are called on the interface.  That
way, the AST is loaded as needed rather than all at once.  Added a few
tests to verify behavior (however, they are not comprehensive; there are
a lot more cases that should be verified in the future).

Change-Id: I64038f9668942a67e1f1b7dceac6c7dbed2d46d7
Signed-off-by: Matthew Bastien <mbastien@blackberry.com>