Bug 484900: Remove CDI
(This is WIP still. It works correctly, i.e. you can run/debug. I would
not expect the changes as is to be ready for merging. Any early feedback
is welcome, including ditching this whole change if it is not suitable.)

Notes about the changes:

- I have removed a lot of CDI at this point, but not all of it.
- I have done no refactoring or moving classes, only removals. The few
insertions were the minimum necessary to compile and work
- JTAG and Remote debugging/launching has not been tested at all

Things left to do:

- Remove most of the public interfaces of CDI
(org.eclipse.cdt.debug.core.model) are still in place, many of them are
part of DSF so its not as simple as delete the whole package
- Activities need to be updated/reviewed
- The plugin.xmls still have references to classes now removed
- LocalCDILaunchDelegate should be removed, for now I have only removed
all the debug parts of it, but most of what is left delegates to
LocalRunLaunchDelegate
- Surely many other things
- Review change made to CApplicationLaunchShortcut, in particular part
marked "TODO: Removing CDI"
- Cleanup/Rename CDI* classes used by DSF
- Other?  

Change-Id: I80274e1a0d77bc7bb00a2afe4babc00c4f7613ae
Signed-off-by: Jonah Graham <jonah@kichwacoders.com>