Bug 484909: Remove duplicate creation of DsfSourceLookupParticipant
The implicit creation of DsfSourceLookupParticipant has been removed
from DsfSourceDisplayAdapter. 

In the GDB case this prevents the duplicate participant from being
created. 

However the PDA example relied on the implicit creation. Therefore,
update the PDA launch to behave as the GDB launch does.

There are two changes that affect other implementers of the DSF (these
are done by DSF-GDB already):
- The implicit DsfSourceLookupParticipant creation will need to be
updated like PDA to add the participant as part of the director's
initializeParticipants()
- To be able to connect the participant to the session, the director can
no longer be created from the plugin.xml (via
DebugPlugin.getDefault().getLaunchManager().newSourceLocator(type)) and
needs to be created explicitly in the code.

Change-Id: I294406845ab91aca2170d90dd77f17d7d751e12f
Signed-off-by: Jonah Graham <jonah@kichwacoders.com>