Bug 484909: Remove duplicate creation of DsfSourceLookupParticipant
The implicit creation of DsfSourceLookupParticipant has been removed
from DsfSourceDisplayAdapter.

In the GDB case this prevents the duplicate participant from being
created.

However the PDA example relied on the implicit creation. Therefore,
update the PDA launch to behave as the GDB launch does.

There is a change that affects other implementers of the DSF (done by
DSF-GDB already):
- The formerly implicit DsfSourceLookupParticipant creation will need to
be updated like PDA to add the participant as part of the director's
initializeParticipants()

Change-Id: I294406845ab91aca2170d90dd77f17d7d751e12f
Signed-off-by: Jonah Graham <jonah@kichwacoders.com>