Update patch set 5
Patch Set 5:

I decided to redo this patch, to push the explicit casting down to where it happens at runtime. This means lots of function signatures do end up changing, but after reviewing how the class is used I came to the conclusion that was better. The class is internal and none of the affected methods are referenced outside of DisassemblyDocument itself (even though they are public).

Note that there is no compile time checking that only AddressRangePosition are added to the Document Managed Positions lists, but there never has been. If some client of DisassemblyDocument did violate the API and put in a non-AddressRangePosition an exception will happen on the retrieval. With this version of the patch the down cast is at the same spot in the code as the down cast actually happens at runtime, rather than being at some unknown point earlier.

Patch-set: 5
Reviewer: Gerrit User 27769 <27769@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0