Bug 484993 - Add support for IRemoteConnection launch targets.
Creates a new core plug-in that bridges the gap between ILaunchTarget
and IRemoteConnection.

Also adds unique id's to ILaunchTarget that are separate from the
user visible name. This is to allow for unique names in o.e.remote
which include the connection type and the connection name. The launch
target type is for all o.e.remote targets.

Change-Id: I08d4c3534fc0947f946bfcdec3f13df04fc4ec98