Bug 484994 - Migrate Mars Arduino support to Neon.
This brings the master branch up to date with everything we've done
in the 8.8 branch. Adjusts for the new ILaunchTarget and targeted
launch delegate.

Also has a start at making more things common in the new build system.
With three extenders of it now in CDT, i.e. Qt, CMake, and Arduino,
it's obvious we can make things more common and make new extenders
less work.

Also undoes some of the work I've done to get Arduino onto the new
build system. Will need to redo it as we do the new common stuff.

Change-Id: I51ce768e0fc60e29c16b05567bd9802d64e33779