Update patch set 11
Patch Set 11: Code-Review-1 Verified-1

(7 comments)

Nicely done!
A couple of in-lines with the major one being simply to add the dependency to easyMock to the target file.  It is available in Orbit already.

Also, I'm getting failures on the SourceLookup tests with GDB 7.1 to GDB 7.4.  GDB 7.5 and up all pass.  Do you also get failures?

Patch-set: 11
Label: Verified=-1
Label: Code-Review=-1