Bug 484894: Update generification of DisassemblyDocument
This fix follows on from af49d7701ab719a7176ad941bc91b956931c8133 and
redoes the generification changes brought on by the AbstractDocument
changes for Neon M4.

The earlier fix (af49d77) made a copy of the list to work around the
casting problem, however many of the methods that work on the positions
expected access to the real list, not a copy.

For example, consider addModelPositionFirst(), it gets the
CATEGORY_MODEL list and then adds to it. If getAddressRangePositions
returned a copy, then the wrong list would be updated.

Change-Id: I36ca589ba748b66541c632182aceaf0b0b64aea4
Signed-off-by: Jonah Graham <jonah@kichwacoders.com>