Update patch set 14
Patch Set 14:

Arg! the gcc on the build machine is very old (4.3.4, released Aug 2009, but 4.3 branch dates back to March 2008) and does not support the flag for testing with old gdb (-gstrict-dwarf, added in late 2009). The gcc is old enough  that it probably just works with older gdbs. However I am not going to test that.

I'll later provide a further patch that can does not regress the tests.

BTW if the Makefile fails to build in the tests, that does not cause the tests to fail en masse, but instead the failures are deferred until the unbuilt exe is not found. In this case the core file is generated after the SourceLookup*.exes so the core file and some of the source lookup tests show as failing.

Patch-set: 14
Label: Verified=0