Update patch set 9
Patch Set 9:

> Jonah, patch 5 is the best approach.
 > 
 > I even would go further with adding deprecated annotation for
 > DisassemblyDocument.addPositionLast(String, Position) to signal
 > that it will be removed in favor of DisassemblyDocument.addPositionLast(String,
 > AddressRangePosition)


I have added the deprecation notice, I also added it to computeIndexInPositionListLast 

Note that there is still a lot of methods (some inherited from AbstractDocument) that allows non AddressRangePositions to be added to a list that later assumes it was a AddressRangePosition (by down casting to it). Of course this was always the case and the generification of AbstractDocument simply exposes a problem that was always there.

Patch-set: 9