Update patch set 2
Patch Set 2:

> My question is unanswered but I haven't found any problem related
 > to this.

 > My question is unanswered but I haven't found any problem related
 > to this.

Sorry Marc-Andre, this issue got neglected due to my other work. Actually, I don't think the pipe gets closed in this way. I'm not a n expert in this matter, but as far as I understand this we read and write on the same file descriptor. Now, if we close it on the PTYOutputStream (write side) it gets closed on the PTYInputStream (read side) too. Subsequently, we don't close the write side and just send EOT instead. In this way we can still read the output of the program until it terminates. The termination will close the file descriptor.

Actually, you can still write to the console (i.e. the PTYOutputStream), but the running Programm should not be able to read this, as it is beyond EOT.

Patch-set: 2
Label: Verified=0