Update patch set 14
Patch Set 14:

Hello, I checked the changes done by you, they seem fine, I made a minor change to correct a mistake after which the patch worked for me. Regarding the icon, you will be able to see it with a gdb 7.10 and atleast a Haswell processor.

Patch-set: 14
Label: Verified=0