Bug 472765: change from easymock to mockito
Due to improvements in target platform in 12843ef it is now possible to
use mockito instead of easymock. This commit does that switch.

Note because of Bug 334540 we need to explicitly require the org.hamcrest
bundle so that it resolves. From the bug, Matthias Sohn says in Comment 1:
"I would just use Require-Bundle to get around this."

Change-Id: Ib49a767c826aa04f4daf1de0bf1b25a4f8ae07ba
Signed-off-by: Jonah Graham <jonah@kichwacoders.com>