Update patch set 3
Patch Set 3:

> (1 comment)
 > 
 > Looks good. Just some extra things that could be cleaned-up:
 > In MIVarListChildrenInfo, this can be removed:
 > else if (val instanceof MIList)
 > In MIStackListLocalsInfo this can be removed:
 > else if (value instanceof MITuple)
 > 
 > and the relevant comments in the files.

Wow, I would not have found those.

However, it is not totally clear what is really Apple-specific is what is not.  Considering the low benefit of removing this to the potential risk of breaking something on linux or windows, I suggest we leave it.

Patch-set: 3
Reviewer: Gerrit User 457 <457@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0