Update patch set 3
Patch Set 3:

I forget now, but couldn't we avoid all the changes to GdbLaunch and GDBBackend by having the new launch delegate write the gdb path in the launch configuration directly?  Then everything else stays the same.

The Backend service was originally created to allow people to create their own service and 'massage' all these settings of the launch config.  Specifically, it was to support running GDB itself on a remote machine and modify all the paths accordingly.

But for that to work, the rest of the code must go through the backend process.  I'm concerned that giving access to all those fields through GdbLaunch will open the door to by-passing the backend service.  Extenders could achieve the same result as extending the backend service by now extending GdbLaunch, but why require them to make this change if we can continue to use the backend service?

Patch-set: 3
Label: Code-Review=0