Update patch set 4
Patch Set 4:

Sorry missed your out of source comments. But this one in particular: "I forget now, but couldn't we avoid all the changes to GdbLaunch and GDBBackend by having the new launch delegate write the gdb path in the launch configuration directly?  Then everything else stays the same."

No, you are not allowed to write to a launch configuration outside of the UI. Launch delegates should never change launch configurations. That's what launches are for.

Patch-set: 4