Update patch set 20
Patch Set 20:

(1 comment)

Hello, I did the changes for the preference page and the hardware trace checking u suggested in the previous comments and thank you for them. I am now waiting for the rest of the changes that may have to follow for the comment in GdbSessionAdapter. If you could point me those then I can submit the next patchset. BTW did u get to see the icons for hardware trace ? You need a minimum Haswell machine with a GDB that has it configured.

Patch-set: 20