Update patch set 4
Patch Set 4:

> No, you are not allowed to write to a launch configuration outside
 > of the UI. Launch delegates should never change launch
 > configurations. That's what launches are for.

Are you saying launches are allowed to change the launch config?  Then couldn't the delegate set the gdb path in the launch, and the launch then change the launch config?  Sounds the same as having the delegate change the launch config, so I must have mis-understood your comment.

Patch-set: 4
Label: Verified=0