Qt - Include tern deps in build. Clean up some old files.
The tern deps were in node_modules which is gitignored. Created a
Makefile to copy those deps to a new directory and we check that in.

Also removed some files that were used to get ANTLR which we aren't
using. And cleaned out the commented out reload button from the
plugin.xml.

Change-Id: I09a5ecb298bbdd9272f9b4a7540d525edb55c4fa