stand-alone debugger: Use rcp repository for creating the products
With this option, the build implicitly verifies that it would also
be possible to install the product (and its plugins) from that
repository.

This means that it provides some basic sanity check for plugin
dependencies.

For example, if you remove org.tukaani.xz from the product contents,
now it fails wheareas before, it would have built a broken product
silently.

See
https://eclipse.org/tycho/sitedocs/tycho-p2/tycho-p2-director-plugin/materialize-products-mojo.html#source

Change-Id: I77382bb99b95784b18892cabe545c039d568ff01
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>