stand-alone debugger: Use rcp repository for creating the products
With this option, the build implicitly verifies that it would also
be possible to install the product (and its plugins) from that
repository.

This means that it provides some basic sanity check for plugin
dependencies.

For example, this would have uncovered the missing dependency to
launchbar.core fixed here: https://git.eclipse.org/r/65572

See
https://eclipse.org/tycho/sitedocs/tycho-p2/tycho-p2-director-plugin/materialize-products-mojo.html#source

Change-Id: I77382bb99b95784b18892cabe545c039d568ff01
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>