Update patch set 6
Patch Set 6:

Ok, I've looked at the new version.  What bothers me is that we are picking certain variables like CC, but omitting others.  For example, a user may need to set all of CC, LD, AS, AR, NM, etc.. to accommodate a correct build.  The use of the user field is then needed as it is today.  I think the right thing to do that would alleviate your concern about intuitiveness in the UI would be to have an Environment variables section in the configure tool settings that followed the standard table format with buttons that one sees in Environment properties pages (e.g. see Local C Program run configuration).  This would make it obvious to the end-user and wouldn't stick the user with only a few variables.  The .autotools file could either have an environment section or add to the user section as it does today.  What do you think?

Patch-set: 6
Label: Verified=0