Bug 473107 - [TERMINALS] Show-In menu type depends on project type
Introduced an ISourceProvider to check if external executables are
configured. Renamed the Terminal context menu "Show In" to "Show in
Local Terminal". Extended the visibleWhen check so that the "Show in
Local Terminal" menu is only visible in the context menu if an external
executable is configured. Otherwise it is possible to open the Terminal
via the inclusion of the command in the default "Show In" menu.

Change-Id: I618b2352a8973d6591cf9e673dc4eb01aeec072e
Signed-off-by: Dirk Fauth <dirk.fauth@googlemail.com>