Update patch set 2
Patch Set 2: Code-Review+1

This looks good to me! (And it's slightly simpler than what I described in comment 6, which is nice. For some reason I thought we'd need to track these flags on a per-template parmeter basis, but thinking more about it I don't see why we'd need to.)

There was one issue: after re-wrapping the argument into a pack expansion, instantiateArguments() needs to clear the packExpanded flag, so that if a subsequent argument is _not_ a pack, it doesn't incorrectly get wrapped on the next iteration. I extended the test case to cover this situation, and a couple others.

I also made a small change to resolveTemplateTypeParameter() - I believe the meaning is the same as before, just written a bit more concisely.

Patch-set: 2
Reviewer: Gerrit User 973 <973@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Code-Review=+1