Bug 456116 - Cannot use class fields in GdbLaunchDelegate
The protected method GdbLaunchDelegate.cleanupLaunch() has been removed.
It has been replaced with GdbLaunchDelegate.cleanupLaunch(ILaunch).

The complexity about not using a class field in GdbLaunchDelegate is that
throughout the process of launching a session, we must be ready to cleanup
the GdbLaunch object.  The problem is that the platform does not provide
that launch object in the methods preLaunchCheck() and finalLaunchCheck(),
so we needed to store the launch object in a field.

This patch delays the call to GdbLaunch.initialize() so that we don't need
to cleanup the launch object if preLaunchCheck() and finalLaunchCheck()
abort the launch.  One hurdle was that we still needed to create the DsfSession
when constructing GdbLaunch, so still needed to clean it up.  This patch adds
a check in GdbLaunch itself, when the launch is removed, to cleanup any
lingering DsfSession automatically.

I've tested every path that aborts the launch in GdbLaunchDelegate and
confirmed that things are properly disposed of in every case.

Change-Id: Ie2981a843917b464f1785a477871073227e027c3