Bug 478740: Cleanup debugger tab
In preparation for new fields that will be added to the debugger tab
to support hardware reverse debugging, this patch does some cleanup.

The idea is to decouple the gdb and gdbinit boxes which require three
columns, from the other options which require one or two columns.  This
will make modifying options simpler and more versatile.

Change-Id: I11909ed72237128f354dbf05ab3ba1de062aad2f