Update patch set 1
Patch Set 1:

> Can you please clarify why we need this change? Making the flag in
 > question final is a good thing to do but there is no way to change
 > it after the delegate is created, right?

I'm trying to cleanup the fact that we used fields in our launch delegates which is not correct.  You had brought to my attention that launch delegates are shared between different launches and therefore that we could not use fields to store information about a particular launch.

When trying to remove such fields, I realized that for fields that don't change and are set by the constructor, we are ok.  Changing such fields after construction would be bad for the same reason that the delegate is shared between launches.

In this case, the field in question is set when constucting different delegates that inherit from AbstractCLaunchDelegate2.  Making it officially final guarantees we won't have issues by someone changing it by mistake in the future.

Patch-set: 1
Reviewer: Gerrit User 457 <457@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0