Bug 478740: Cleanup debugger tab
This patch cleans up the debugger tab and puts the new ReverseMode
dropdown box on the line as the reverse checkbox.

The patch also decouples the gdb and gdbinit boxes which require three
columns, from the other options which require one or two columns.  This
will make modifying options simpler and more versatile.

Change-Id: I11909ed72237128f354dbf05ab3ba1de062aad2f