Bug 489398 - Support memory spaces in IExpressionDMAdress
When an address is resolved from an expression, the resulting address
may carry a memory space provided by the backend to qualify
the address to a memory space.

 The current interface for the memory space service (e.g.
IMemorySapce2) can be used to parse the memory space contained in the
expression.

This update adds:
1) A new memory space field to IExpressionDMAddress
2) The use of the memory space service to attempt to resolve the memory
space.

Note: if there is no memory space service, the memory space defaults to
an empty string, which does not affect the default behaviour i.e.
when memory spaces are not used.

Change-Id: Idfe5669b26f84ee4e3e78f96f229ced75e6ec5c3