Bug 489501 - [memory] Add API to resolve the default memory space id
When using memory spaces, a backend may sometimes qualify an address
to specify the memory space to which the address belongs.  However, to
minimize impacts, the memory space qualifier may not always be present.
In such cases, the address is implied to belong to the "default" memory
space.

DSF-GDB currently has no means to know what is the "default" memory
space to apply in such cases.

This change introduces a new API to request the default memory space
from the memory service.

Change-Id: Icea87543b8529756aee92c6d2af95fb936f5fb9c