Bug 489501 - [memory] Add API to resolve the default memory space id
When using memory spaces, a back end may sometimes qualify an address
to specify the memory space this address belongs to,
however to minimize impacts the memory space may not always be present
e.g. as it will use it to refer to the default memory space.

However DSF-GDB has no means to know what is the default memory space
to apply to e.g. and address resulting from an expression.

This change introduces a new API to request the memory service for the
default memory space.

Change-Id: Icea87543b8529756aee92c6d2af95fb936f5fb9c