Update patch set 1
Patch Set 1:

Found another (double) case: in the console, in the save action. The confirmation dialog, for the user to confirm the overwriting of existing file, would (at least some times) not pop-up, because the shell was null. In the same method, any I/O error while saving the file would also be missing its error dialog...

Patch-set: 1