Bug 303808: Support for full GDB console
This patch provides a full GDB console in the standard Console view.
The full console is disabled for Windows due to PTY limitations.

The new console will be triggered automatically when using GDB 7.12 (or
its pre-release) and running on Linux.

Known issues with this patch:
- Cloning the console view breaks the console
- Pinning the console view is not supported
- Interrupting execution in all-stop mode fails with the new console
- Pagination causes some events to be lost (all-stop only)

The all-stop issues cause many failures in our JUnit test suite.

Change-Id: Iee6ef5228ca17bd829eb743cb41a142afe6714dc