Update patch set 1
Patch Set 1: Code-Review+1

I have reviewed this +1. There is some public API that is not strictly CDI that is left unreferenced by this and it points to the need to refactor launch component.

The entire hierarchy of org.eclipse.cdt.launch.ui.CMainTab should be removed (this is distince from DSF's CMainTab)

Some of the classes in the hierarchy of CMainTab are unused/unreferenced in CDT, however RemoteCMainTab is still used: It is still the case that for remote launch the Main Tab is different when doing Run and Debug. When doing Run the org.eclipse.cdt.launch.remote.tabs.RemoteCMainTab class is used and when doing Debug the org.eclipse.cdt.launch.remote.tabs.RemoteCDSFMainTab class is used.

Patch-set: 1
Label: Code-Review=+1