Update patch set 2
Patch Set 2:

MIkhail's contribution was delayed because we agreed (in comments in the bug I believe) on how to improve the solution.  Mikhail had no time at the moment though.

It would be good if you looked at the discussion to see if your support addresses those points, and if you even agree with them.

Patch-set: 2