Update patch set 2
Patch Set 2: Code-Review-1

The new action type is intended to run an existing GDB script file. See https://git.eclipse.org/r/#/c/52520/. 
This patch adds an action type that allows users to run a sequence of commands they enter in the dialog. It is not exactly the same because some users want to use scripts with the IDE and the CLI gdb.

Patch-set: 2
Label: Code-Review=-1