Bug 489516 - Traditional mem., choose visibility of the cross ref information
This facility is dependent of the issues below.

Bug 489512 - Show local variables in the traditional memory render
Bug 489513 - Show registers in the traditional memory render

Allow the user to choose visibility of the available cross reference
information e.g. Variables, Registers

In addition to the preferences at a instance level, a global preference
is introduced to disable the resolution of additional information
altogether.

Change-Id: I9f8b9b612ff288b369ba6fca3402c0c911aa5f48