Bug 489677 - Fix remote launch race condition
After uploading an executable, we run chmod on the target to make the
fiel actually executable (in order to work around a bug in the sftp
implementation).  However, we don't wait until the remote shell process
running the chmod is done before continuing.  The result is that
gdbserver tries to execute the file before it's executable.

This patch ensures that we wait for the shell to be done before carrying
on.

I first tried to use p.waitFor(), but its implementation in
HostShellProcessAdapter seems to wait for at least two seconds.  It
would add some unnecessary delay to the launch.

Change-Id: If0e916b0648cfa6a0ddd355417c7dad5fce41787
Signed-off-by: Simon Marchi <simon.marchi@polymtl.ca>