Update patch set 9
Patch Set 9:

> What is your feel about the value of this change?  I also don't
 > like to have two participants, but I hesitate at risking breaking
 > other DSF users for no reason beyond cleanup.

This is a pre-requisite for correct operation of "Bug 472765: Use gdb's "set substitute-path from to"". We must not have a DsfSourceLookupParticipant in that case because it does not understand the set substitute-path, GdbSourceLookupParticipant does understand it. If you have both you get unexpected behaviour.

As for the other reasons:

I don't know the magnitude of effect on a large project. It is more serious on source not found as all the lookups are done twice. (Note that in 8.8 there were three lookups done, but commit 28b1830acf942216ae99d4d167b6c3b2133aaec6 change 63288 already cleaned up one of them.)

The other issue is that if you have the implicit participant added by the DsfSourceDisplayAdapter then the adapter cannot be used if you don't want to have DsfSourceLookupParticipant (e.g. you only want a descendent of it, the case at the top.)

As for breaking other DSF users. To break another user they would have to not be using DsfSourceLookupDirector too as that is what creates the "original" of the duplicate pair.

Patch-set: 9
Label: Verified=0