Update patch set 10
Patch Set 10:

Sorry for the comment churn here.

Unfortunately one of my statements above is no longer true. I had forgotten that I removed the dependency from change 52068 to this change. 

The rest of the arguments are still true, but perhaps that is outweighed by the change being a cleanup one only.

Patch-set: 10