Update patch set 29
Patch Set 29:

> Marc, Thanks for the review.
 > 
 > As for the complicated case, one thing I had not considered was
 > that the user can simply order the source lookups if they are in
 > this difficult situation. I will update the loop soon. I think
 > there are still some cases where things can break, but they become
 > more and more contrived and all are certainly not handled correctly
 > today, so we are moving in the correct direction.
 > 
 > BTW Frankly I am at a loss how to clearly explain (aka document)
 > the case where someone legitimately has two path mappings, one
 > handled by backend and one not.

I think your example was pretty good at making me understand that we cannot fix all cases.  If the user wants to break things, they can.  Personally, I don't think we need to worry about this anymore.  I would vote for doing what is easier to understand in the code, as long as it does work with the actual use-cases you are addressing.

Just let me know when you have posted a patchset that satisfies you and I'll commit it.

Thanks for your hard work, this was not a trivial change to make, and you handled it beautifully.

Patch-set: 29