Update patch set 13
Patch Set 13: Code-Review+2

I think we  should go ahead with this one so that the new file structure is in place for M6.

I think there are probably some oversights in this change that it would be nice to fix, so I would like to take the time to do a more thorough review, but this will have to wait until after M6.  I can do the review post-commit.

Patch-set: 13
Label: Code-Review=+2