Bug 484900: Replace LocalCDILaunchDelegate with LocalRunLaunchDelegate
The launch delegate id "org.eclipse.cdt.cdi.launch.localCLaunch" was
being used for both the Run launch and the CDI debug launch.  To keep
supporting the CDI debug launch, we still had to keep that delegate
pointing to the class LocalCDILaunchDelegate.

Now that we have remove the CDI debug launch, we no longer need to keep
pointing to LocalCDILaunchDelegate, and can use the
"org.eclipse.cdt.cdi.launch.localCLaunch" on for Run and pointing to
LocalRunLaunchDelegate.

Some more details can be found in Bug 464636.

Change-Id: I36c2adc043d29a5a11121059470505c6526c068b