Update patch set 5
Patch Set 5: -Code-Review

> I was just reviewing the class hierarchy of org.eclipse.cdt.launch.ui.CMainTab.
 > It really should go at some point because it is part of what causes
 > different looking UI in Run and Debug. However, as CDT's
 > org.eclipse.cdt.launch.remote.tabs.RemoteCMainTab and TCF's
 > org.eclipse.tcf.internal.cdt.ui.launch.RemoteCMainTab still are
 > extenders I recommend we simply deprecate this entire hierarchy for
 > now and progress with removing the plugin.xml stuff.

Thanks Jonah for bringing up this inconsistency in the UI.
I have pushed review 68799 which move our remote launch to use the DSFMainTab instead of the old CDI one.

I will do another commit to cleanup the CMainTab hierarchy.  We can work with the TCF folks to adapt them, althoug the change is a one-line to change RemoteCMainTab to RemoteCDSFMainTab.

Patch-set: 5