Update patch set 8
Patch Set 8: Code-Review+2

Finally, the build passed.
This change is pretty big in terms of API impact.  However, if we don't do it, we are setting ourselves up for a big problem when we need to use GDB's breakpoint id which is already not always a integer.  I'll push this.  I will also document the broken APIs.

Patch-set: 8
Label: Code-Review=+2